+++ packages/yourloops/components/patient-data/patient-data.tsx
@@ -27,7 +27,7 @@
 
 import React, { type FunctionComponent, useEffect, useRef, useState } from 'react'
 import { PatientNavBarMemoized as PatientNavBar } from '../header-bars/patient-nav-bar'
-import { Navigate, Route, Routes } from 'react-router-dom'
+import { Navigate, Route, Routes, useParams } from 'react-router-dom'
 import { AppUserRoute } from '../../models/enums/routes.enum'
 import { PrintPDFDialog } from '../pdf/print-pdf-dialog'
 import { PatientDashboard } from '../dashboard-widgets/patient-dashboard'
@@ -51,15 +51,19 @@ import { useAuth } from '../../lib/auth'
 import { DeviceView } from '../../pages/patient-view/device/device-view'
 import { setPageTitle } from '../../lib/utils'
 import { TargetAndAlertsView } from '../../pages/patient-view/target-and-alerts/target-and-alerts-view'
-import { useSelectedTeamContext } from '../../lib/selected-team/selected-team.provider'
 import TeamUtils from '../../lib/team/team.util'
+import { Patient } from '../../lib/patient/models/patient.model'
 
-export const PatientData: FunctionComponent = () => {
+interface PatientDataProps {
+  patient: Patient
+}
+
+export const PatientData: FunctionComponent<PatientDataProps> = ({ patient }: PatientDataProps) => {
   const alert = useAlert()
   const theme = useTheme()
   const { t } = useTranslation()
   const patientIdForWhichDataHasBeenFetched = useRef(null)
-  const { selectedTeam } = useSelectedTeamContext()
+  const { teamId } = useParams()
 
   const {
     bgPrefs,
@@ -79,10 +83,9 @@ export const PatientData: FunctionComponent = () => {
     refreshData,
     refreshingData,
     trendsDate,
-    patient,
     timePrefs,
     updateChartPrefs
-  } = usePatientData()
+  } = usePatientData({ patient })
   const {
     showMessageCreation,
     showMessageThread,
@@ -227,7 +230,7 @@ export const PatientData: FunctionComponent = () => {
                     }
                   />
                   {
-                    user.isUserHcp() && !TeamUtils.isPrivate(selectedTeam) &&
+                    user.isUserHcp() && !TeamUtils.isPrivate(teamId) &&
                     <Route
                       path={AppUserRoute.TargetAndAlerts}
                       element={
@@ -237,7 +240,8 @@ export const PatientData: FunctionComponent = () => {
                       }
                     />
                   }
-                  <Route path="*" element={<Navigate to={AppUserRoute.Dashboard} replace />} />
+                  <Route path="/" element={<Navigate to={AppUserRoute.Dashboard} replace />} />
+                  <Route path="*" element={<Navigate to={AppUserRoute.NotFound} replace />} />
                 </Routes>
                 {showPdfDialog &&
                   <PrintPDFDialog