+++ packages/yourloops/test/integration/pages/verify-email/verify-email.common.spec.tsx
@@ -28,11 +28,10 @@
 import * as auth0Mock from '@auth0/auth0-react'
 import { getAccessTokenWithPopupMock, logoutMock } from '../../mock/auth0.hook.mock'
 import { renderPage } from '../../utils/render'
-import { screen, waitFor, within } from '@testing-library/react'
-import { checkFooterForUserNotLoggedIn } from '../../assert/footer.assert'
-import userEvent from '@testing-library/user-event'
+import { waitFor } from '@testing-library/react'
 import { AUTH0_ERROR_EMAIL_NOT_VERIFIED } from '../../../../lib/auth/models/auth0-error.model'
 import { AppRoute } from '../../../../models/enums/routes.enum'
+import { testVerifyEmail } from '../../use-cases/email-verification'
 
 describe('Verify email page', () => {
   it('should display a description of the email verification process with options', async () => {
@@ -51,49 +50,6 @@ describe('Verify email page', () => {
       expect(router.state.location.pathname).toEqual(AppRoute.VerifyEmail)
     })
 
-    checkFooterForUserNotLoggedIn(true)
-
-    const pageHeader = within(screen.getByTestId('verify-email-header'))
-    expect(pageHeader.getByLabelText('YourLoops Logo')).toBeVisible()
-    const headerLogoutButton = pageHeader.getByText('Logout')
-    expect(headerLogoutButton).toBeEnabled()
-
-    await userEvent.click(headerLogoutButton)
-    expect(logoutMock).toHaveBeenCalledTimes(1)
-
-    const pageContent = within(screen.getByTestId('verify-email-content'))
-    expect(pageContent.getByLabelText('YourLoops Logo')).toBeVisible()
-    expect(pageContent.getByText('Verify your email address')).toBeVisible()
-    expect(pageContent.getByText('To use YourLoops, you need to confirm your email address by clicking on the link we sent you. This helps to keep your account secure.')).toBeVisible()
-
-    const contactSupportSection = pageContent.getByTestId('verify-email-details-2')
-    expect(contactSupportSection).toHaveTextContent('No mail in your inbox or spam folder? Contact the customer support.')
-    const contactSupportButton = within(contactSupportSection).getByRole('button', { name: 'Contact the customer support' })
-    expect(contactSupportButton).toBeVisible()
-
-    await userEvent.click(contactSupportButton)
-    expect(window.open).toHaveBeenCalledTimes(1)
-
-    const createNewAccountSection = pageContent.getByTestId('verify-email-details-3')
-    expect(createNewAccountSection).toHaveTextContent('Wrong address? If you mistyped your email when signing up, logout and create a new account.')
-    const logoutAndCreateNewAccountButton = within(createNewAccountSection).getByRole('button', { name: 'logout and create a new account' })
-    expect(logoutAndCreateNewAccountButton).toBeVisible()
-
-    await userEvent.click(logoutAndCreateNewAccountButton)
-    expect(logoutMock).toHaveBeenCalledTimes(2)
-
-    const logoutSection = pageContent.getByTestId('verify-email-details-4')
-    expect(logoutSection).toHaveTextContent('Do you want to login with another address? Try to login again.')
-    const tryToLoginAgainButton = within(logoutSection).getByRole('button', { name: 'Try to login again' })
-    expect(tryToLoginAgainButton).toBeVisible()
-
-    await userEvent.click(tryToLoginAgainButton)
-    expect(logoutMock).toHaveBeenCalledTimes(3)
-
-    const continueButton = pageContent.getByText('Continue')
-    expect(continueButton).toBeEnabled()
-
-    await userEvent.click(continueButton)
-    expect(screen.getByTestId('alert-snackbar')).toHaveTextContent('You have not verified your email.')
+    await testVerifyEmail()
   })
 })