+++ packages/yourloops/test/unit/lib/team/team.util.test.ts
@@ -27,7 +27,6 @@
 
 import { buildTeam } from '../../common/utils'
 import TeamUtils from '../../../../lib/team/team.util'
-import { TeamType } from '../../../../lib/team/models/enums/team-type.enum'
 
 describe('TeamUtils', () => {
   describe('sortTeamsByName', () => {
@@ -44,11 +43,8 @@ describe('TeamUtils', () => {
 
   describe('isPrivate', () => {
     it('should check whether the given team is the private team', () => {
-      const team = buildTeam('teamId', [])
-      const privateTeam = buildTeam('private', undefined, undefined, TeamType.private)
-
-      expect(TeamUtils.isPrivate(team)).toEqual(false)
-      expect(TeamUtils.isPrivate(privateTeam)).toEqual(true)
+      expect(TeamUtils.isPrivate('teamId')).toEqual(false)
+      expect(TeamUtils.isPrivate('private')).toEqual(true)
     })
   })
 })