+++ test/locales/translations.test.js
@@ -280,7 +280,11 @@ const ignoredTransKeyForYourLoops = [
   'total-declared-carbs',
   'total-estimated-carbs',
   'time-loop-tooltip',
-  'tooltip-total-derived-carbs'
+  'tooltip-total-derived-carbs',
+  'verify-email-error-title',
+  'tooltip-total-derived-carbs',
+  'sensor-warmup',
+  'sensor-warmup-session-end'
 ]
 const ignoredTransKeyInYourLoopsFiles = [
   'yourloops|${s}',